Updated calcMedian method to return the smaller of the two middle values, rather than the average of the two, as was directed to do in class

+++ Registrar.h
@@ -40,18 +40,21 @@ public:
   //  Calc Stat functions
   //    student stats
   float CalcMeanStudentWait();
-  float CalcMedianStudentWait();
+  int   CalcMedianStudentWait();
   int   LongestStudentWait();
   int   NumStudentsOver10Minutes();
   //    window stats
-  int   MeanWindowIdleTime();
+  float MeanWindowIdleTime();
   int   LongestWindowIdleTime();
   int   NumWindowsOver5Minutes();
 
 
+
+
+
 private:
   void MakeIdleList();
-
+  
   GenQueue<Student> studentQueue;
   Window* windowArray;
   int numWindows;
